pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "hikerapi>=1.7.3",
    "gunicorn>=21.2.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
